# types: New benchmarks + new (Any).Union algorithm