--- a/types/types_bench_test.go
+++ b/types/types_bench_test.go
@@ -38,3 +38,57 @@ func generateType(n int) Type {
 	}
 	return NewObject(static, nil)
 }
+
+func generateTypeWithPrefix(n int, prefix string) Type {
+	static := make([]*StaticProperty, n)
+	for i := 0; i < n; i++ {
+		static[i] = NewStaticProperty(prefix+fmt.Sprint(i), S)
+	}
+	return NewObject(static, nil)
+}
+
+func BenchmarkAnyMergeOne(b *testing.B) {
+	sizes := []int{100, 500, 1000, 5000, 10000}
+	for _, size := range sizes {
+		anyA := Any(make([]Type, 0, size))
+		for i := 0; i < size; i++ {
+			tpe := generateType(i)
+			anyA = append(anyA, tpe)
+		}
+		tpeB := N
+		b.ResetTimer()
+		b.Run(fmt.Sprint(size), func(b *testing.B) {
+			result := anyA.Merge(tpeB)
+			if len(result) != len(anyA)+1 {
+				b.Fatalf("Expected length of merged result to be: %d, got: %d", len(anyA)+1, len(result))
+			}
+		})
+	}
+}
+
+// Build up 2x Any type lists of unique and different types, then Union merge.
+func BenchmarkAnyUnionAllUniqueTypes(b *testing.B) {
+	sizes := []int{100, 250, 500, 1000, 2500}
+	for _, sizeA := range sizes {
+		for _, sizeB := range sizes {
+			anyA := Any(make([]Type, 0, sizeA))
+			for i := 0; i < sizeA; i++ {
+				tpe := generateType(i)
+				anyA = append(anyA, tpe)
+			}
+			anyB := Any(make([]Type, 0, sizeB))
+			for i := 0; i < sizeB; i++ {
+				tpe := generateTypeWithPrefix(i, "B-")
+				anyB = append(anyB, tpe)
+			}
+			b.ResetTimer()
+			b.Run(fmt.Sprintf("%dx%d", sizeA, sizeB), func(b *testing.B) {
+				resultA2B := anyA.Union(anyB)
+				// Expect length to be A + B - 1, because the `object` type is present in both Any type sets.
+				if len(resultA2B) != (len(anyA) + len(anyB) - 1) {
+					b.Fatalf("Expected length of unioned result to be: %d, got: %d", len(anyA)+len(anyB), len(resultA2B))
+				}
+			})
+		}
+	}
+}