@philipaconrad 2023-09-19
@johanfylling I ran a quick version of your [optimization idea](https://github.com/open-policy-agent/opa/pull/6228#pullrequestreview-1630777341) on my laptop this morning, and on the `BenchmarkCompileDynamicPolicy` benchmark that we've been using as a proxy for how "solved" issue #5216 is, that optimization nets a *10x or better improvement* across *all* metrics (runs, ns/op, B/op, allocs/op).