@philipaconrad 2023-09-19
:information_source:  Squashed down the changes for this PR, will file a separate PR for the optimization idea @johanfylling proposed.