@anderseknert 2023-09-19
It would be useful if this included a few examples of policy, or relatable operations that OPA performs, which benefits from this. It's all rather abstract as it is, even if I'm sure the change is great :)