@philipaconrad 2023-09-19
@anderseknert I was a bit worried about that as well when I first filed the PR. :sweat_smile: This PR *does* slightly help on one of the partial evaluation benchmarks (`InliningFullScan`), where it provides a fairly reliable 1-1.5% speedup (30x runs each for `main` vs PR).

```
name                       old time/op    new time/op    delta
InliningFullScan/1000-8      12.7ms ± 5%    12.5ms ± 6%  -1.54%  (p=0.034 n=30+28)
InliningFullScan/10000-8      152ms ± 3%     152ms ± 6%    ~     (p=0.594 n=28+30)
InliningFullScan/300000-8     4.17s ± 3%     4.12s ± 2%  -1.20%  (p=0.000 n=28+28)

name                       old alloc/op   new alloc/op   delta
InliningFullScan/1000-8      5.91MB ± 0%    5.91MB ± 0%    ~     (p=0.269 n=30+27)
InliningFullScan/10000-8     59.4MB ± 0%    59.4MB ± 0%    ~     (p=0.767 n=30+30)
InliningFullScan/300000-8    1.78GB ± 0%    1.78GB ± 0%    ~     (p=0.424 n=30+30)

name                       old allocs/op  new allocs/op  delta
InliningFullScan/1000-8        179k ± 0%      179k ± 0%    ~     (all equal)
InliningFullScan/10000-8      1.79M ± 0%     1.79M ± 0%    ~     (p=0.461 n=30+30)
InliningFullScan/300000-8     53.7M ± 0%     53.7M ± 0%    ~     (p=0.970 n=30+30)
```

From what I've seen, any part of OPA that does type checking should expect to see slightly better average-case performance from this PR, so I'd expect this to hold for most things that involve the `compile.Compiler` type, and *possibly* for schema-related type checking as well, although we don't have benchmarks for that system at this time.